inference/benchmark/requirements.txt
awscli==1.33.33
beartype==0.19.0
boto3==1.34.144
filelock===3.15.4
inflect==4.1.0
levenshtein_rs==0.1.7
matplotlib==3.9.1
orjson==3.10.6
pandas==2.2.2
pause==0.3
pydantic==2.8.2
requests==2.32.3
sox==1.4.1
text_unidecode==1.3
tqdm==4.66.4
websocket-client==1.8.0
